analytics-service/requirements.txt
fastapi
uvicorn
aio-pika
pika
psycopg2-binary
sqlalchemy
pydantic
jsonschema
river
scikit-learn
xgboost
redis
python-dotenv
pyyaml
tenacity
prometheus-client